refactor(backend): add get_current_user, build_context to utils.py
