+++ website/presentation/routes/posts_routes.py
@@ -1,8 +1,8 @@
 from flask import Blueprint, render_template, redirect, url_for, flash, request, jsonify
-from flask_login import current_user, login_required
+from flask_login import login_required
 
 from website import limiter
-from website.config import Config
+from website.utils import get_current_user, build_context
 from website.presentation.forms import CreatePostForm, CommentForm
 from website.presentation.middlewares import admin_required, token_required
 from website.application.services import PostService, CommentService
@@ -15,21 +15,6 @@ posts_bp = Blueprint(
 )
 
 post_service = PostService()
-
-
-def get_current_user():
-    return current_user if current_user.is_authenticated else None
-
-
-def build_context(user, active_page=""):
-    is_admin = bool(user and user.role.name == "ADMIN")
-    return {
-        "is_admin": is_admin,
-        "avatar_url": user.avatar_url if user else "",
-        "token": Config.SECRET_KEY if is_admin else "",
-        "theme": user.theme.value if user else "system",
-        "active_page": active_page,
-    }
 
 
 @posts_bp.route("/", methods=["GET"])
